# apps/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Triathlon API",
    "version": "0.2.0",
    "description": "One public HTTP interface for all Triathlon clients. Error responses share the Problem Details envelope with a stable domain code."
  },
  "components": {
    "securitySchemes": {
      "bearerKey": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/sprints": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "plannedStart": {
                    "type": "string"
                  },
                  "plannedEnd": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "projectId",
                    "name",
                    "plannedStart",
                    "plannedEnd",
                    "state",
                    "activatedAt",
                    "completedAt",
                    "createdBy",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "plannedStart": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "plannedEnd": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "state": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "planned"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "active"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "completed"
                          ]
                        }
                      ]
                    },
                    "activatedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "completedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "projectId",
                      "name",
                      "plannedStart",
                      "plannedEnd",
                      "state",
                      "activatedAt",
                      "completedAt",
                      "createdBy",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "projectId": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "plannedStart": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "plannedEnd": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "state": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "planned"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "active"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "completed"
                            ]
                          }
                        ]
                      },
                      "activatedAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "completedAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "createdBy": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/sprints/{sprintId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sprintId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sprint",
                    "members"
                  ],
                  "properties": {
                    "sprint": {
                      "type": "object",
                      "required": [
                        "id",
                        "projectId",
                        "name",
                        "plannedStart",
                        "plannedEnd",
                        "state",
                        "activatedAt",
                        "completedAt",
                        "createdBy",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "projectId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "plannedStart": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "plannedEnd": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "state": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "planned"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "active"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "completed"
                              ]
                            }
                          ]
                        },
                        "activatedAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "completedAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "createdBy": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    },
                    "members": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "ticketId",
                          "addedAt",
                          "addedBy",
                          "removedAt"
                        ],
                        "properties": {
                          "ticketId": {
                            "type": "string"
                          },
                          "addedAt": {
                            "type": "string"
                          },
                          "addedBy": {
                            "type": "string"
                          },
                          "removedAt": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sprintId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/sprints/{sprintId}/activate": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sprintId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "projectId",
                    "name",
                    "plannedStart",
                    "plannedEnd",
                    "state",
                    "activatedAt",
                    "completedAt",
                    "createdBy",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "plannedStart": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "plannedEnd": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "state": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "planned"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "active"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "completed"
                          ]
                        }
                      ]
                    },
                    "activatedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "completedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/sprints/{sprintId}/complete": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sprintId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "projectId",
                    "name",
                    "plannedStart",
                    "plannedEnd",
                    "state",
                    "activatedAt",
                    "completedAt",
                    "createdBy",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "projectId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "plannedStart": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "plannedEnd": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "state": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "planned"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "active"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "completed"
                          ]
                        }
                      ]
                    },
                    "activatedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "completedAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/sprints/{sprintId}/members": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ticketIds"
                ],
                "properties": {
                  "ticketIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sprintId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sprint",
                    "members"
                  ],
                  "properties": {
                    "sprint": {
                      "type": "object",
                      "required": [
                        "id",
                        "projectId",
                        "name",
                        "plannedStart",
                        "plannedEnd",
                        "state",
                        "activatedAt",
                        "completedAt",
                        "createdBy",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "projectId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "plannedStart": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "plannedEnd": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "state": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "planned"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "active"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "completed"
                              ]
                            }
                          ]
                        },
                        "activatedAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "completedAt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "createdBy": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    },
                    "members": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "ticketId",
                          "addedAt",
                          "addedBy",
                          "removedAt"
                        ],
                        "properties": {
                          "ticketId": {
                            "type": "string"
                          },
                          "addedAt": {
                            "type": "string"
                          },
                          "addedBy": {
                            "type": "string"
                          },
                          "removedAt": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{projectId}/sprints/{sprintId}/snapshot": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "projectId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "sprintId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1"
    }
  ]
}